third_party/blink/web_tests/virtual/disable-css-grid-lanes-layout/external/wpt/css/css-grid/parsing/grid-template-rows-intrinsic-auto-repeat-computed-implicit-track.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Property grid-template-rows value '1px repeat(auto-fill, auto) 3px' computes to '10px 1px 0px 3px'
  assert_equals: expected "10px 1px 0px 3px" but got "none"
[FAIL] Property grid-template-rows value '1px repeat(auto-fit, auto) 3px' computes to '10px 1px 0px 3px'
  assert_equals: expected "10px 1px 0px 3px" but got "none"
[FAIL] Property grid-template-rows value '1px [a] repeat(auto-fill, min-content max-content) [b] 4px' computes to '10px 1px [a] 0px 0px [b] 4px'
  assert_equals: expected "10px 1px [a] 0px 0px [b] 4px" but got "none"
[FAIL] Property grid-template-rows value '1px [a] repeat(auto-fit, min-content max-content) [b] 4px' computes to '10px 1px [a] 0px 0px [b] 4px'
  assert_equals: expected "10px 1px [a] 0px 0px [b] 4px" but got "none"
[FAIL] Property grid-template-rows value '1px [a] repeat(auto-fill, [b] fit-content(200px) [c]) [d] 3px' computes to '10px 1px [a b] 0px [c d] 3px'
  assert_equals: expected "10px 1px [a b] 0px [c d] 3px" but got "none"
[FAIL] Property grid-template-rows value '1px [a] repeat(auto-fit, [b] fit-content(200px) [c]) [d] 3px' computes to '10px 1px [a b] 0px [c d] 3px'
  assert_equals: expected "10px 1px [a b] 0px [c d] 3px" but got "none"
[FAIL] Property grid-template-rows value '[a] 1px repeat(auto-fill, auto [b] auto) 4px [d]' computes to '10px [a] 1px 0px [b] 0px 4px [d]'
  assert_equals: expected "10px [a] 1px 0px [b] 0px 4px [d]" but got "none"
[FAIL] Property grid-template-rows value '[a] 1px repeat(auto-fit, auto [b] auto) 4px [d]' computes to '10px [a] 1px 0px [b] 0px 4px [d]'
  assert_equals: expected "10px [a] 1px 0px [b] 0px 4px [d]" but got "none"
[FAIL] Property grid-template-rows value '100% [a] repeat(auto-fill, [b] min-content [c]) [d] 300%' computes to '10px 1px [a b] 0px [c d] 3px'
  assert_equals: expected "10px 1px [a b] 0px [c d] 3px" but got "none"
[FAIL] Property grid-template-rows value '100% [a] repeat(auto-fit, [b] min-content [c]) [d] 300%' computes to '10px 1px [a b] 0px [c d] 3px'
  assert_equals: expected "10px 1px [a b] 0px [c d] 3px" but got "none"
[FAIL] Property grid-template-rows value '[a] 1em repeat(auto-fill, max-content [b] auto) 4em [d]' computes to '10px [a] 1px 0px [b] 0px 4px [d]'
  assert_equals: expected "10px [a] 1px 0px [b] 0px 4px [d]" but got "none"
[FAIL] Property grid-template-rows value '[a] 1em repeat(auto-fit, max-content [b] auto) 4em [d]' computes to '10px [a] 1px 0px [b] 0px 4px [d]'
  assert_equals: expected "10px [a] 1px 0px [b] 0px 4px [d]" but got "none"
Harness: the test ran to completion.
